Ho mergiato dev nella mia brach e ho applicato un po di modifiche,Cosi funziona per bene e ho usato il runner per riempire il db con i municipi e le province

+++ src/main/java/team6/epicenergyspa/exceptions/NotFoundException.java
@@ -5,7 +5,7 @@ public class NotFoundException extends RuntimeException {
         super(message);
     }
 
-    public NotFoundException(int id) {
+    public NotFoundException(Long id) {
         super(id + " non trovato!");
     }
 }
